+++ src/components/NavBar/_NavBar.scss
