+++ src/components/_index.scss
@@ -1,2 +1,11 @@
 @import './BtnToTop/BtnTotop';
 @import './AboutResume/AboutResume';
+@import './ContactForm/ContactForm';
+@import './Footer/Footer';
+@import './Hero/Hero';
+@import './LikeToDo/LikeToDo';
+@import './ListWorksSkills/ListWorksSkills';
+@import './NavBar/NavBar';
+@import './Projects/Projects';
+@import './Services/Services';
+@import './Skills/Skills';
