+++ src/index.scss
@@ -26,31 +26,22 @@
 // Contact Me
 @import './scss/views/ContactMe.scss';
 // Contact Me (Form)
-@import './scss/components/ContactForm';
 
 // Hero
-@import './scss/components/Hero';
 
 // NavBar
-@import './scss/components/NavBar';
 
 // LikeToDo (Cards)
-@import './scss/components/LikeToDo';
 
 // ListWorksSkills
-@import './scss/components/ListWorksSkills';
 
 // Skills
-@import './scss/components/Skills';
 
 // Projects
-@import './scss/components/Projects';
 
 // Services
-@import './scss/components/Services';
 
 // Fotter
-@import './scss/components/Footer';
 
 // Components
 @import './components/index';
